third_party/WebKit/LayoutTests/fast/canvas/canvas-text-space-characters-expected.txt

PASS testStr1Width == testStr2Width is true
PASS successfullyParsed is true

TEST COMPLETE
